Fix data race on search.
The VPTree type was not thread-safe during search.  The `tau` variable
was shared across all instances.  This patch removes it from the struct
and adds it as a parameter to search() instead.

I've also a test for use with `go test -race -cpu=4``.  Without the fix,
this test reports a data race.