Fix duplicate distanceMetric computation (thanks @vseledkin!)
It's not obvious at first that this change is correct because the elements get
shuffled around. But `items[median]` in this instance refers to the same item we
used to compute `pivotDist` before the points were partitioned, so we don't need
to do the (potentially expensive) calculation of the distance metric twice.

Closes #2.